+++ src/vec3.rs
@@ -2,20 +2,20 @@ use std::fmt::Debug;
 
 #[derive(Copy, Clone, Debug)]
 pub struct Vec3 {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
 }
 
 pub type Point = Vec3;
 pub type Color = Vec3;
 
 impl Vec3 {
-    pub fn length_squared(&self) -> f32 {
+    pub fn length_squared(&self) -> f64 {
         self.dot(self)
     }
 
-    pub fn length(&self) -> f32 {
+    pub fn length(&self) -> f64 {
         self.length_squared().sqrt()
     }
 
@@ -23,7 +23,7 @@ impl Vec3 {
         println!("{} {} {}", self.x, self.y, self.z);
     }
 
-    pub fn dot(&self, other: &Self) -> f32 {
+    pub fn dot(&self, other: &Self) -> f64 {
         self.x * other.x + self.y * other.y + self.z * other.z
     }
 
@@ -39,7 +39,7 @@ impl Vec3 {
         *self / self.length()
     }
 
-    pub fn new(x: f32, y: f32, z: f32) -> Self {
+    pub fn new(x: f64, y: f64, z: f64) -> Self {
         Self { x, y, z }
     }
 }
@@ -69,7 +69,7 @@ impl Neg for Vec3 {
 }
 
 impl Index<u8> for Vec3 {
-    type Output = f32;
+    type Output = f64;
 
     fn index(&self, index: u8) -> &Self::Output {
         match index {
@@ -158,10 +158,10 @@ impl MulAssign for Vec3 {
     }
 }
 
-impl Mul<f32> for Vec3 {
+impl Mul<f64> for Vec3 {
     type Output = Self;
 
-    fn mul(self, scalar: f32) -> Self {
+    fn mul(self, scalar: f64) -> Self {
         Self {
             x: self.x * scalar,
             y: self.y * scalar,
@@ -170,7 +170,7 @@ impl Mul<f32> for Vec3 {
     }
 }
 
-impl Mul<Vec3> for f32 {
+impl Mul<Vec3> for f64 {
     type Output = Vec3;
 
     fn mul(self, vec: Vec3) -> Vec3 {
@@ -204,10 +204,10 @@ impl DivAssign for Vec3 {
     }
 }
 
-impl Div<f32> for Vec3 {
+impl Div<f64> for Vec3 {
     type Output = Self;
 
-    fn div(self, scalar: f32) -> Self {
+    fn div(self, scalar: f64) -> Self {
         Self {
             x: self.x / scalar,
             y: self.y / scalar,
